Fix `tests/racket/conditions005` by adding `ignore`.
As pointed out by gallais, `IO` actions are not allowed to silently
discard their result anymore. This means we need to explicitly ignore
the results of `for threads $ \t => threadWait t`, which is safe to do
since waiting on a thread produces an `IO ()`, so `for` produces a
`List (IO ())`.