Use `blodwen-stringbytelen` instead of C's `strlen` (fixes #1260)
Use the already existing (both Chez and Racket) support function to
accurately determine the UTF-8 byte count of a string, as the C function
(unsurprisingly) stops counting at NUL characters.

Currently the only usage of `stringByteLength` is in the implementation
`TTC String`, which leads to string truncation when a string containing
NUL bytes is serialised to a .ttc file.