Don't calculate blocked metavariable set
With the new representation of postponed problems, this is now either
not worth doing, or much more expensive than just having another crack
at the problem anyway. So let's not calculate it.

This makes the libraries built a couple of seconds quicker for me, and I
have one example where a program that previously took minutes now type
checks in seconds...

Might revisit this later, if it turns out to be a potential source of
optimisations, but it'll need to be done an entirely different way.