[ fix ] Make case-split respect indentation for impossible cases.
When case-splitting on something which is impossible, e.g. in an
implementation of `Uninhabited`, the `showImpossible` function from
`Idris/IDEMode/CaseSplit.idr` previously did not take indentation into
account, often resulting in generating invalid code which needed to be
re-indented manually. This commit adds a function which calculates the
indentation, and modifies `showImpossible` to take indentation into
account on code generation. It also adds tests so that this behaviour is
not accidentally removed in the future.