[ test, refactor ] Run tests in parallel using base primitives.
The previous test-runner implementation relied on `Future` from
`contrib` and spawned a new future for each test instead of re-using the
threads (expected behaviour in a pool). This commit removes the
dependency on `Future` (since `contrib` libs are generally considered
experimental) in favour of using the built-in concurrency primitives. It
also re-uses threads for running the tests in each TestPool, by sending
tests and receiving results over `Channel`s. No behaviour in the tests
is altered; they are still observed to work correctly and run in
parallel.