Tidy up applyToStack/continueNF
It's best if they don't use 'quote' (which actually is a potential
problem anyway, if quotes somehow end up nested).

applyToStack actually only worked properly for NTCon/NDCon/NPrim,
although in practice that was fine since it was only ever called in
those contexts. This fixes it so it properly continues evaluation, which
means we have a chance of caching results in more cases.