Add CBV mode to the evaluator
I'm not sure how useful this is going to be, but I set it as the default
for evaluation at the REPL, since it can save duplicating work.

It's not a good choice for evaluation during type checking, since we
want to avoid evaluation when we don't need it. Also when we resume
suspended unification problems, the context may be different due to
solved metavariables, so even if we evaluate early we might still have
to evaluate again.