Issue #1572
Added "Profile" as a flag that can be used when building a project.  Since, if you try to use :exec within a repl, the profile flag is ignored by the Chez backend, without this flag, and there is no "profile" setting available within the "ipkg" file, there is not way to turn profiling on with the Chez backend.