Reorganise tests so that they have correct CG
We shouldn't have tests requiring a CG in the fundamental language
feature tests, because we don't know if all the CGs will behave the
same. So, if we want to be sure they do, we should have individual tests
for each CG. Now the tests are arranged so that any one that uses a CG
is behind the check that the required CG exists.
I'll shortly add a check to the test suite so that we don't accidentally
add tests that use a CG where we shouldn't.