[ contrib ] Add `Queue`s and `BufferedChannel`s.
This commit adds shared, non-blocking, and MTSafe Queues and
BufferedChannels. The queues are implemented based on the book "Purely
Functional Data Structures" by Chris Okasaki. The buffered channels are
then implemented on top of queues. Both Queues and BCs also come with
tests in `tests/contrib` to verify that their behaviour works as
expected.

The repository/package this commit is based on can be found here:
https://github.com/CodingCellist/idris2-buffered-channels