[ base ] Fix #1552. N.B.: Changes Channel semantics in Chez cg.
After much discussion and debugging with @zseri, this seems to be the
simplest _fully working_ implementation of Channels under the
Chez-Scheme backend. However, since the main cause of trouble was that
`channelPut` should block until the value had been received, this is no
longer true in the Chez backend. Instead, `channelPut` now only blocks
if there is already a value in the Channel which hasn't been received
yet. This is different to the Racket Channel behaviour, but at least it
consistently works.