Data.IOArray.Prims.prim__newUninitArray
`prim__newArray` requires an element to fill in the array.

Consider we want to implement an `IOArray` without `Maybe` wrapper
(like in #1677).

To implement `fromList` we need to create an array. When the list
is not empty we can implement call `prim__newArray` taking a first
element from the list.  But when the list is empty, we need to this
`prim__newEmptyArray` function.

Additionally this function might be needed for more efficient
implementation of functions like `fromList`: no need to initialize
array if the array will be immediately overwritten.