Make foreign calls collect_safe by default
__collect_safe in Chez allows C functions to run without worrying about
scheme memory management. The only reason a function wouldn't be collect
safe is if it has access to some scheme managed memory, which with the
FFI as it stands, is only Buffers, and we can spot those.
This means we don't need to have a special C calling convention just for
Chez, and we don't need to expect users to flag potentially blocking
foreign functions.
By the way: __collect_safe only works on Chez >= 9.5.1 anyway, so using
earlier versions may lead to blocking in concurrent programs. We should
document this!