Rename C__collect_safe call convention to C__blocking
The idea is this. RefC codegen might need to support threading.
The obvious way to suport threading is to implement GIL (as Python
does).

To make GIL useful, we need to unlock it sometimes. As in Python,
we can unlock GIL when we are performing blocking native calls.

Thus `C__blocking` convention can be used in RefC codegen to emit
code similar to `C` convention with addition of unlocking/locking
GIL.