Enable incremental compilation on Windows.
Update Windows launch script.

Add back the current directory to source-directories (not needed for
this, but it's the default and it's better to keep load working as usual,
there's no risk as the new dir has priority).

Copy libs instead of linking when testing (linking has bad interactions
with admin privileges).