FAQ/CONTRIBUTION updates
We had a lot of things missing here, and the contributing guidelines
were extremely out of date and reflected the state when I first made the
Idris 2 repo public. I've updated both to reflect the current state of
the way we work, and to give better guidelines about what will be most
helpful.