Work a bit harder to run delayed elaborators
We used to choose a particular order for delayed elaborators (that is,
e.g., things that needed some more information to resolve an ambiguity,
or case blocks waiting for full type information) and only run them once
for the sake of performance. The evidence is that this is not necessary,
and performance issues with ambiguity are elsewhere, so now we keep
going with the delayed elaborators.

They're still given priority ordering (case blocks last) because this
gives a handy heuristic for better error messages. In particular, if a
case block can't elaborate, it's often due to an ambiguity resolution
failure, and we want to report that first.

I will be interested to know what, if any, effect this has particularly
on performance. If you're seeing this message as a result of a git
bisect: Hi! I've been expecting you!