Mark prim__system C__collect_safe
It is a blocking call, so it should be marked `__collect_safe`.
Could not check if it really helps (or if the issue exists) because
on my mac laptop `system` seems to hold the global lock.  Checked
with this program:

```
#include <pthread.h>
#include <stdlib.h>

void* t(void *p) {
    system("echo tt; sleep 1; echo TT");
    return 0;
}

#define VERIFY(cond) do { if (!(cond)) abort(); } while (0)

int main() {
    pthread_t pt;
    VERIFY(pthread_create(&pt, 0, t, 0) == 0);
    system("echo mm; sleep 1; echo MM");
    void* r;
    VERIFY(pthread_join(pt, &r) == 0);
    return 0;
}
```

which outputs:

```
mm
MM
tt
TT
```

Chez `__collect_safe` function cannot accept `string` arguments, so this diff:
* adds `strdup` C function
* uses `strdup` to convert `String` to a pointer
* passes this pointer to `prim__system` function

The alternative to this PR is to revert #1620: when `system` for
Chez is implemented natively by Chez, it does some chez-specific
low-level machinery which probably handles blocking correctly.

I prefer this PR because:
* `strdup` costs nothing compared to the process spawn
* it allows keeping `system` implementation in different codegens same