[ refc ] Alternative arbitrary precision integers implementation with libbf
Since this is not a feature needed by many people, I tried to minimize the changes to the code so the user need to download libbf himself, and has to make decision during bootstraping and cannot change mind when running refc codegen. The rational and usage is documented in `refc.rst`.