Implement generic interpolation
String interpolation can now take values other than strings as long as
they implement the `Interpolation` interface. The `Interpolation`
interface has 1 method `interpolate : a -> String` that converts
a value of the given type into a string for interpolation. By default
only one instance is given for `String` and the method is the identity
function.

Interpolation is now desugared as calls to `concatInterp` which performs
concatenation of two values that both conform to `Interpolation`. The
values are converted to strings first and then concatenated with
classical string interpolation