Allow linear variables to solve auto implicits
It's currently a little limited, because of the potential variation in
elaboration order we have to be careful to pick the most recently bound
potential result, and we have to be able to solve it immediately (so, if
determining arguments are still to be resolved, you're out of luck).

Perhaps I can make it less limited, but I think that'll take more
drastic changes to the elaborator, and I'm not going to do that for a
while. Consider it an experiment for now.