[ build ] Simplify bootstrap
Delete a whole lot of things from bootstrap-stage2.sh that doesn't
seem to be needed.

Reasoning: LD_LYBRARY_PATH is already handled by the wrapper
scripts. IDRIS2_PATH is not exported and is not propagated anywhere in
the Makefile.

It causes no visible harm when testing it on Linux using:

`make SCHEME=chez-scheme clean bootstrap test`