Generalize `doParse` to work in any underlying `MonadToken`
The expectation of a `MonadToken` is that it implements two effects:
  * Supply the `nextToken`
  * Allow checkpointing of token stream state

The old `doParse` is recovered using `State (List (WithBounds tok))`,
but we can also create new parsers that use e.g. an unpacked byte buffer
and an index into it. This is especially useful for the JS backend, where
just transforming the input into a cons-list representation can lead to a
stack overflow from as little as 40k elements, in my experience.