[ re #2010 ] Implement feedback; fixes multiple oneline `case-of`.
The previous version of PR #2010 didn't account for multiple `case-of`
statements, spotting only the first `of` when it should have been
spotting the last `of`. By using the fish operator (`<><`) to construct
a SnocList and matching on that, the code correctly handles the multiple
oneline `case-of` statements, and it is more readable!

With thanks to @gallais for spotting this and suggesting how to start
fixing things.