Fix for incremental builds
When we're inlining, since we only have the current module, we might
encounter names we don't have definitions for. In this case, we need to
leave the arity alone, since the module it originates from will have set
it correctly.

This is tricky to test in isolation, and ideally needs something in CI
for a full incremental build (which would have caught the problem!). I
don't have time right now, but let's at least see if this breaks
anything else for now.